Fixes as per AM comments

+++ check-copyright.sh
@@ -1,12 +1,12 @@
 #!/bin/sh
 
-set -eu
-
 # Copyright: 2013, Vasudev Kamath <kamathvasudev@gmail.com>
 # License: MIT
 
+set -eu
+
 if [ $# -lt 1 ]; then
-    echo "Usage: $0 package[s]"
+    echo "Usage: $0 package[s]" >&2
     exit 2
 fi
 
@@ -19,38 +19,38 @@ CURDIR=$(pwd)
 # further use.
 create_tmp_workspace() {
     local pkgpath="$1"
-    tdir=$(mktemp -d --tmpdir=/tmp)
-    cp $pkgpath $tdir
-    echo $tdir
+    tdir=$(mktemp -d)
+    cp $pkgpath "$tdir"
+    echo "$tdir"
 }
 
 # Function first checks if given file exists or not and if it exists
 # it checks the extension of input is deb or not.
-validate_package_name(){
+is_valid_file(){
     local package="$1"
-    
-    [ -f $package ] || false
+
     ext=$(basename $package | sed 's/.*\.//')
-    [ "$ext" = "deb" ] || false
+    [ "$ext" = "deb" -a -f "$package" ] || echo "false"
 }
 
-# Function extracts the .deb archive then checks the debian-binary
-# version if its not 2.0 Currently 2.0 is only considered as supported
+# Function extracts the .deb archive then checks the content of
+# debian-binary version. Currently 2.0 is only considered as supported
 # format.
 #
 # It extracts data and control compressed archives to respective
 # directory.
 extract_archive(){
     archive="$1"
-    ar_extracted=$(ar xv $(basename $archive) | awk '{ print $3}')
-    for file in $ar_extracted; do
-	if [ "$file" != "debian-binary" ] ; then
-	    dirpart=$(echo $file | sed 's/\.tar.*//')
-	    mkdir $dirpart
-	    tar -C "$dirpart" -xaf $file
-	else
-	    [ $(cat "$file") != "2.0" ] && (echo "Format of Debian archive is not supported!" && exit 2)
-	fi
+    ar xv $(basename $archive) 2>&1 > /dev/null
+    for file in $(ar t $(basename $archive)); do
+        if [ "$file" != "debian-binary" ] ; then
+            dirpart=$(echo $file | sed 's/\.tar.*//')
+            mkdir "$dirpart"
+            tar -C "$dirpart" -xaf "$file"
+        else
+            [ $(cat "$file") != "2.0" ] && (echo "Format of Debian \
+            archive is not supported!" >&2 && exit 2)
+        fi
     done
 }
 
@@ -60,10 +60,10 @@ verify_copyright(){
     local pkg="$1"
     local copyright_file=$(find data -name copyright -print)
 
-    if [ "$copyright_file" = "data/usr/share/doc/$package/copyright" ]; then
-	echo "copyright for $pkg [Found]"
+    if [ -f data/usr/share/doc/$package/copyright ]; then
+        echo "copyright for $pkg [Found]"
     else
-	echo "copright for $pkg [Not Found]" 
+        echo "copright for $pkg [Not Found]"
     fi
 }
 
@@ -78,24 +78,27 @@ print_version() {
 # Lets iterate over all command line arguments.
 for debpkg in "$@"; do
     # check if we got all Deb packages
-    validate_package_name $debpkg
-    
-    # create work space for the script
-    tdir=$(create_tmp_workspace $debpkg)
+    if [ -z $(is_valid_file "$debpkg") ]; then
 
-    # go to the workspace
-    cd "$tdir"
-    
-    # process archive (.deb package)
-    extract_archive $debpkg
+        # create work space for the script
+        tdir=$(create_tmp_workspace $debpkg)
 
-    package=$(basename $debpkg|awk -F"_" '{ print $1}')
-    echo --------------------------------------------------
-    verify_copyright $package
-    print_version $package
-    echo --------------------------------------------------
+        # go to the workspace
+        cd "$tdir"
 
-    cd $CURDIR
-    rm -rf $tdir
+        # process archive (.deb package)
+        extract_archive "$debpkg"
+
+        package=$(basename $debpkg|awk -F"_" '{ print $1}')
+        echo --------------------------------------------------
+        verify_copyright $package
+        print_version $package
+        echo --------------------------------------------------
+
+        cd $CURDIR
+        rm -rf $tdir
+    else
+        echo "Given file is not a valid deb package" >&2
+        exit 2
+    fi
 done
-
